Refactor Model Viewer to work with Next app router
Currently I've done a first pass at adapting the search query hooks to work with next/navigation, and created a STAC context, provider, and hook to allow loading the catalog, but not connected a page in to the app router.
